compile_to_file makes more sense in module
